+++ packages/core/src/languages/utils/mangler-expressions/index.ts
@@ -1,1 +1,2 @@
+export { default as NestedGroupExpression } from "./nested-group.class";
 export { default as SingleGroupManglerExpression } from "./single-group.class";